+++ obj.py
@@ -1,6 +1,6 @@
 # Andrea Abril Palencia Gutierrez, 18198
-# SR4: Flat Shading --- Graficas por computadora, seccion 20
-# 27/07/2020 - 04/08/2020
+# SR5: Textures --- Graficas por computadora, seccion 20
+# 04/08/2020 - 10/08/2020
 
 # carga de archivo obj
 class Obj(object):
